requirements.txt
diff-match-patch==20121119
Django==1.11.17
django-angular==1.1.4
django-ckeditor==5.3.1
django-import-export==0.5.1
django-js-asset==0.1.1
easy-thumbnails==2.4.2
et-xmlfile==1.0.1
jdcal==1.3
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2017.2
PyYAML==3.12
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
